Schema part of SQLite database url is `sqlite://`
The current implementation assumes the schema part of a SQLite database url to be `sqlite:///`, it swallows the root path slash when a user config the `c.JupyterHub.db_url` item using an absolute path. In this case, `_check_db_path` will get a wrong relative path and report file missing.