Created a two-layer process start up for starting up the spawner cmd
An intermediate process is now used to create the spawner's child process.
This allows for PAM sessions to properly be created.
A variable called spawner.proc_state is used to pass the pid and return code
of the process back to the main process. It has three states: None, multiprocessing.queues.Queue, or int
None means not process has been started.
When a Queue is started the pid is passed as the first value, and then the return code is passed when the
child process exits.
If spawner.proc_state is an int, it means the program has finished and the int is the returncode.