Add options to validate HTML before sending to the page.
While obviously this is a lot of extra work; This is mostly to get the
conversation started; and think as whether this should be set for – for
example – testing.

That would allow to catch earlier miss-matched tags; or issues in
templates.

I'm not quite sure AdvancedHTMLParser is the best for that; but I
haven't found a better library.