WIP: Add GET /servers API
This is the start of adding a GET /servers API which attempts
to be as close to the GET /users admin API with just the servers
resource portion of that response without needing to query the
users table in the database.

This is eventually something that is going to be used when
culling idle servers without querying users to improve performance
in environments with a large number of users (tens of thousands).

Obviously work needs to be done with translating the orm.Spawner
object to a spawner.Spawner object to build the response, docs,
more tests, etc. This is a start to get feedback and show it works.

Part of issue 2954