Run test_get_servers early
This just moves the servers API tests earlier in the
module so that user API tests which spawn servers do
not interfere with the results of the servers API tests.
Note that cleanup_after fixture is set to autouse=True
but either it doesn't work or that fixture is buggy in
that it only deletes records with active spawners but
GET /servers returns *all* spawners, not just active ones.