Fix test_get_servers/users
There are two servers when test_api runs - one for the user
created by the mock app instance and an admin created in an
earlier test, so test_get_servers is updated.

As a result of running test_get_servers earlier than
test_get_users the non-admin user resource has activity so
that test needs to be updated.