Return user name and server id from GET /servers
This fleshes out the TODOs in the code so that we convert the
orm.Spawner instance to a proper spawner.Spawner instance and
set the user.User object on it so that the server resource
returned in GET /servers is the same as what's in GET /users.

In addition, for GET /servers specifically we return the id
and user name so that things like cull_idle_servers can list
(and eventually delete) servers without needing to interact
with the users API if it's not culling users.

Docs will come in a follow up change.