[WIP] Start work on the DELETE /servers/{id} API
This starts the DELETE /servers/{id} API based on the
existing DELETE /users/{name}/servers/{server_name} API.
TODOs are inline for testing and DRYing up some of that
code, plus docs are obviously needed.

There is also a question of whether or not we want to do
a separate route to stop a server versus delete it to keep
this code a bit less complicated.

Part of issue 2954