Add happy path test for DELETE /servers/{id}
This adds a test which creates a named server for the non-admin
user and then deletes it as admin. This exposes a problem in the
API where the non-admin user should be able to delete their own
server but the DELETE /servers/{id} API requires the server id
which only comes from GET /servers which is admin-only.