Change cull_idle to use GET /servers if not culling users
This changes the cull_idle_servers script such that if users
are not being culled (the default) and the GET /servers API
is available it uses that to get servers directly rather than
go through GET /users which could have a lot more records if
servers are being culled but users are not.

In our deployment we have ~50K users but only a couple hundred
servers and since cull_idle_servers runs on startup it makes
restarting our single jupyterhub pod instance quite slow:

  [I 2020-02-27 13:32:07.798 JupyterHub log:174] \
  200 GET /hub/api/users (cull-idle@127.0.0.1) 1222100.09ms

There is a bit of refactoring in this change to have consistent
"log names" and to be graceful about GET /servers not being
available.

Closes issue #2954