Remove check for allowed_users for users created via config
adding users via config anywhere makes them allowed

previously, this was _required_, so that it was always true for working config,
but config which allowed some users but declared others in groups or roles was forbidden.

Now, declaring a user anywhere _ensures_ the user is allowed rather than _enforcing_ it.